requirements.txt
Flask
datetime
uuid
Flask-SQLAlchemy
PyJWT
werkzeug
Flask-Migrate
